Closing #551 with this.

Squashed commit of the following:

commit 11f904349a76a68d6bb6deaf76c14e8c675b1753
Author: Stefan Profanter <git@s.profanter.me>
Date:   Thu Mar 3 16:06:41 2016 +0100

    Caching install dir

commit fce442443993c4f78089000b81d293c7b64ee794
Merge: 8eed598 559b88e
Author: Stefan Profanter <git@s.profanter.me>
Date:   Thu Mar 3 15:29:52 2016 +0100

    Merge remote-tracking branch 'upstream/master' and resolve conflicts

commit 8eed598e1283d54420848f4311b4367a9da7c4ba
Author: Stefan Profanter <git@s.profanter.me>
Date:   Thu Mar 3 15:22:09 2016 +0100

    Use valgrind 3.11

commit 2def5858cd68c3e5d52c5a2943909ead079e4cf2
Author: Stefan Profanter <git@s.profanter.me>
Date:   Wed Mar 2 08:32:01 2016 +0100

    Using -j8 flag for make

commit f75ea211d66f677de87fce7c180cd75a24fbec89
Merge: f0170c0 eb9b8e5
Author: Stefan Profanter <git@s.profanter.me>
Date:   Wed Mar 2 08:30:45 2016 +0100

    Merge remote-tracking branch 'upstream/simpify_amalgate'

commit eb9b8e57424d69a406cc661bbd6dbe37da992af0
Author: Stasik0 <github@stasik.com>
Date:   Wed Mar 2 08:19:06 2016 +0100

    trying to make CMake simpler, by removing additional dependencies

commit f0170c06ea9bc11d9a0156464a1c57ce1e3dd244
Author: Stefan Profanter <git@s.profanter.me>
Date:   Wed Mar 2 00:08:58 2016 +0100

    Minor build fixes
    - Add amalgamation to server-static
    - Only switch to gcc-4.8 if gcc is used
    - Build coveralls only if gcc is used
    - cast return value in examples (causes clang to fail otherwise)

commit c1e0e40672f105428754d61ef9a7e5c7dea4956c
Author: Stefan Profanter <git@s.profanter.me>
Date:   Tue Mar 1 15:57:31 2016 +0100

    Cross compile only for gcc

commit eb8e7a26388389576582f50972d80c989742f70f
Author: Stefan Profanter <git@s.profanter.me>
Date:   Tue Mar 1 15:15:46 2016 +0100

    Allow ANALYZE=true builds to fail

commit e659d928bbcdb3fba07ba7b5c3b07fbe83665c4d
Author: Stefan Profanter <git@s.profanter.me>
Date:   Tue Mar 1 15:10:09 2016 +0100

    Using new travis build environment and cppcheck (fix #391 and fix #371)

+++ examples/client_firstSteps.c
@@ -48,5 +48,5 @@ int main(void) {
 
     UA_Client_disconnect(client);
     UA_Client_delete(client);
-    return UA_STATUSCODE_GOOD;
+    return (int) UA_STATUSCODE_GOOD;
 }
